feat(live-region-element): add support for delayMs
